requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
joblib>=1.3
pyyaml>=6.0
pytest>=8.0
